move files assets to appropriate folders
