+++ app/Http/Requests/Api/V1/File/StoreFileRequest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\Api\v1\File;
 
 use Illuminate\Foundation\Http\FormRequest;
